Revert "rating and review screen"
This reverts commit 80b722f82ce2f5a0a70c365b59018b4ac03b7216.

+++ lib/models/product.dart
@@ -71,14 +71,14 @@ late int _countRating;
 }
 
 class Price {
- late double _priceMax;
- late double _priceMin;
- double get priceMax => _priceMax;
- double get priceMin => _priceMin;
+ late int _priceMax;
+ late int _priceMin;
+  int get priceMax => _priceMax;
+  int get priceMin => _priceMin;
 
   Price({
-   required  double priceMax,
-   required  double priceMin}){
+   required  int priceMax,
+   required  int priceMin}){
     _priceMax = priceMax;
     _priceMin = priceMin;
 }
